Update Sailor and Training services for exception handling.
Exception handling has been improved in both SailorService and TrainingService by introduction of SailorNotFoundException and TrainingNotFoundException. These exceptions are now thrown if a sailor or training is not found by their UUID.

+++ src/main/java/com/esaillog/sailor/SailorService.java
@@ -1,15 +1,19 @@
 package com.esaillog.sailor;
 
 import com.esaillog.training.Training;
+import com.esaillog.training.TrainingNotFoundException;
 import com.esaillog.training.TrainingRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class SailorService {
 
     private final SailorRepository sailorRepository;
@@ -19,8 +23,8 @@ public class SailorService {
         return sailorRepository.findAll().stream().map(SailorMapper::toDto).toList();
     }
 
-    public SailorDTO getSailor(String uuid) {
-        return sailorRepository.findById(UUID.fromString(uuid)).map(SailorMapper::toDto).orElseThrow();
+    public SailorDTO getSailor(String uuid) throws SailorNotFoundException {
+        return sailorRepository.findById(UUID.fromString(uuid)).map(SailorMapper::toDto).orElseThrow(SailorNotFoundException::new);
     }
 
     public void addSailor(SailorDTO sailorDTO) {
@@ -28,18 +32,34 @@ public class SailorService {
     }
 
     private List<Training> fetchTrainingsBySailorDTO(SailorDTO sailorDTO) {
-        return sailorDTO.trainings().stream().map(trainingId -> trainingRepository.findById(UUID.fromString(trainingId)).orElseThrow()).toList();
+        return sailorDTO.trainings().stream()
+                        .map(this::findTrainingById)
+                        .filter(Objects::nonNull)
+                        .toList();
     }
 
-    public void updateSailor(String uuid, SailorDTO sailorDTO) {
-        Sailor sailor = sailorRepository.findById(UUID.fromString(uuid)).orElseThrow();
+    private Training findTrainingById(String trainingId) {
+        try {
+            return trainingRepository.findById(UUID.fromString(trainingId)).orElseThrow(TrainingNotFoundException::new);
+        } catch (TrainingNotFoundException e) {
+            log.warn(e.getMessage());
+            return null;
+        }
+    }
+
+    public void updateSailor(String uuid, SailorDTO sailorDTO) throws SailorNotFoundException {
+        Sailor sailor = sailorRepository.findById(UUID.fromString(uuid)).orElseThrow(SailorNotFoundException::new);
         sailor.setFirstName(sailorDTO.firstName());
         sailor.setLastName(sailorDTO.lastName());
         sailor.setEmail(sailorDTO.email());
         sailor.setTrainings(fetchTrainingsBySailorDTO(sailorDTO));
     }
 
-    public void deleteSailor(String uuid) {
-        sailorRepository.deleteById(UUID.fromString(uuid));
+    public void deleteSailor(String uuid) throws SailorNotFoundException {
+        if (sailorRepository.existsById(UUID.fromString(uuid))) {
+            sailorRepository.deleteById(UUID.fromString(uuid));
+        } else {
+            throw new SailorNotFoundException();
+        }
     }
 }
