+++ src/main/java/com/esaillog/training/TrainingService.java
@@ -1,16 +1,20 @@
 package com.esaillog.training;
 
 import com.esaillog.sailor.Sailor;
+import com.esaillog.sailor.SailorNotFoundException;
 import com.esaillog.sailor.SailorRepository;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 @Service
 @AllArgsConstructor
+@Slf4j
 public class TrainingService {
     private final TrainingRepository trainingRepository;
     private final SailorRepository sailorRepository;
@@ -19,8 +23,8 @@ public class TrainingService {
         return trainingRepository.findAll().stream().map(TrainingMapper::toDto).toList();
     }
 
-    public TrainingDTO getTraining(String uuid) {
-        return trainingRepository.findById(UUID.fromString(uuid)).map(TrainingMapper::toDto).orElseThrow();
+    public TrainingDTO getTraining(String uuid) throws TrainingNotFoundException {
+        return trainingRepository.findById(UUID.fromString(uuid)).map(TrainingMapper::toDto).orElseThrow(TrainingNotFoundException::new);
     }
 
     public void addTraining(TrainingDTO trainingDTO) {
@@ -28,11 +32,18 @@ public class TrainingService {
     }
 
     private List<Sailor> fetchSailorsByTrainingDTO(TrainingDTO trainingDTO) {
-        return trainingDTO.sailors().stream().map(sailorId -> sailorRepository.findById(UUID.fromString(sailorId)).orElseThrow()).toList();
+        return trainingDTO.sailors().stream().map(sailorId -> {
+            try {
+                return sailorRepository.findById(UUID.fromString(sailorId)).orElseThrow(SailorNotFoundException::new);
+            } catch (SailorNotFoundException e) {
+                log.warn(e.getMessage());
+                return null;
+            }
+        }).filter(Objects::nonNull).toList();
     }
 
-    public void updateTraining(String uuid, TrainingDTO trainingDTO) {
-        Training training = trainingRepository.findById(UUID.fromString(uuid)).orElseThrow();
+    public void updateTraining(String uuid, TrainingDTO trainingDTO) throws TrainingNotFoundException {
+        Training training = trainingRepository.findById(UUID.fromString(uuid)).orElseThrow(TrainingNotFoundException::new);
         training.setName(trainingDTO.name());
         training.setType(trainingDTO.type());
         training.setHasCertificate(trainingDTO.hasCertificate());
@@ -41,7 +52,11 @@ public class TrainingService {
         training.setSailors(fetchSailorsByTrainingDTO(trainingDTO));
     }
 
-    public void deleteTraining(String uuid) {
-        trainingRepository.deleteById(UUID.fromString(uuid));
+    public void deleteTraining(String uuid) throws TrainingNotFoundException {
+        if(trainingRepository.existsById(UUID.fromString(uuid))) {
+            trainingRepository.deleteById(UUID.fromString(uuid));
+        } else {
+            throw new TrainingNotFoundException();
+        }
     }
 }
